- Moving `to spf13/cobra` library - Ability to wildcard namespace name for single context. - Copy commands to clipboard with hotkeys for Mac. - Adding automatic update functionality. - Status bar with some events. - Colour to execution time when it takes more than 1 second. - Limit number of concurrent getPods request to 3.
