+++ app/gui_test.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 import (
 	"fmt"
@@ -11,91 +11,91 @@ func TestMoveCursor(t *testing.T) {
 	screen := tcell.NewSimulationScreen("")
 	testTable := []struct {
 		name           string
-		frame          Frame
+		frame          InfoFrame
 		moveBy         int
 		expectedCurY   int
 		expectedOffset int
 	}{
 		{
 			name:           "move_down_no_scroll",
-			frame:          Frame{height: 10, cursorY: 0, nsItems: fakeNamespaces(20)},
+			frame:          InfoFrame{height: 10, cursorY: 0, nsItems: fakeNamespaces(20)},
 			moveBy:         5,
 			expectedCurY:   5,
 			expectedOffset: 0,
 		},
 		{
 			name:           "move_down_bottom_border_no_scroll",
-			frame:          Frame{height: 10, cursorY: 8, nsItems: fakeNamespaces(20)},
+			frame:          InfoFrame{height: 10, cursorY: 8, nsItems: fakeNamespaces(20)},
 			moveBy:         1,
 			expectedCurY:   9,
 			expectedOffset: 0,
 		},
 		{
 			name:           "move_down_with_scroll",
-			frame:          Frame{height: 10, cursorY: 8, nsItems: fakeNamespaces(20)},
+			frame:          InfoFrame{height: 10, cursorY: 8, nsItems: fakeNamespaces(20)},
 			moveBy:         2,
 			expectedCurY:   9,
 			expectedOffset: 1,
 		},
 		{
 			name:           "move_down_with_scroll_offset_not_0",
-			frame:          Frame{height: 10, cursorY: 7, scrollYOffset: 3, nsItems: fakeNamespaces(20)},
+			frame:          InfoFrame{height: 10, cursorY: 7, scrollYOffset: 3, nsItems: fakeNamespaces(20)},
 			moveBy:         4,
 			expectedCurY:   9,
 			expectedOffset: 5,
 		},
 		{
 			name:           "move_down_outside_positions_no_scroll",
-			frame:          Frame{height: 10, cursorY: 1, nsItems: fakeNamespaces(6)},
+			frame:          InfoFrame{height: 10, cursorY: 1, nsItems: fakeNamespaces(6)},
 			moveBy:         10,
 			expectedCurY:   5,
 			expectedOffset: 0,
 		},
 		{
 			name:           "move_down_with_scroll_outside_positions",
-			frame:          Frame{height: 10, cursorY: 7, scrollYOffset: 2, nsItems: fakeNamespaces(15)},
+			frame:          InfoFrame{height: 10, cursorY: 7, scrollYOffset: 2, nsItems: fakeNamespaces(15)},
 			moveBy:         10,
 			expectedCurY:   9,
 			expectedOffset: 5,
 		},
 		{
 			name:           "move_down_with_empty_namespaces",
-			frame:          Frame{height: 10, cursorY: 0, scrollYOffset: 0, nsItems: nil},
+			frame:          InfoFrame{height: 10, cursorY: 0, scrollYOffset: 0, nsItems: nil},
 			moveBy:         1,
 			expectedCurY:   0,
 			expectedOffset: 0,
 		},
 		{
 			name:           "move_up_no_scroll",
-			frame:          Frame{height: 10, cursorY: 7, scrollYOffset: 0, nsItems: fakeNamespaces(20)},
+			frame:          InfoFrame{height: 10, cursorY: 7, scrollYOffset: 0, nsItems: fakeNamespaces(20)},
 			moveBy:         -2,
 			expectedCurY:   5,
 			expectedOffset: 0,
 		},
 		{
 			name:           "move_up_to_border_now_scroll",
-			frame:          Frame{height: 10, cursorY: 2, scrollYOffset: 5, nsItems: fakeNamespaces(20)},
+			frame:          InfoFrame{height: 10, cursorY: 2, scrollYOffset: 5, nsItems: fakeNamespaces(20)},
 			moveBy:         -2,
 			expectedCurY:   0,
 			expectedOffset: 5,
 		},
 		{
 			name:           "move_up_with_scroll",
-			frame:          Frame{height: 10, cursorY: 1, scrollYOffset: 5, nsItems: fakeNamespaces(20)},
+			frame:          InfoFrame{height: 10, cursorY: 1, scrollYOffset: 5, nsItems: fakeNamespaces(20)},
 			moveBy:         -3,
 			expectedCurY:   0,
 			expectedOffset: 3,
 		},
 		{
 			name:           "move_up_too-far_with_scroll",
-			frame:          Frame{height: 10, cursorY: 3, scrollYOffset: 5, nsItems: fakeNamespaces(20)},
+			frame:          InfoFrame{height: 10, cursorY: 3, scrollYOffset: 5, nsItems: fakeNamespaces(20)},
 			moveBy:         -20,
 			expectedCurY:   0,
 			expectedOffset: 0,
 		},
 		{
 			name:           "move_up_when_cursor_is_outside_available_positions",
-			frame:          Frame{height: 10, cursorY: 10, scrollYOffset: 15, nsItems: fakeNamespaces(20)},
+			frame:          InfoFrame{height: 10, cursorY: 10, scrollYOffset: 15, nsItems: fakeNamespaces(20)},
 			moveBy:         -15,
 			expectedCurY:   0,
 			expectedOffset: 10,
