+++ buildLocal.sh
@@ -2,6 +2,6 @@
 version=local-$(git describe --tags)
 time=$(date)
 echo "$version"
-go test .
+go test ./...
 go build -o .build/k8ConsoleViewer \
-         -ldflags="-X 'github.com/JLevconoks/k8ConsoleViewer/version.BuildTime=$time' -X 'github.com/JLevconoks/k8ConsoleViewer/version.BuildVersion=$version'" .
+         -ldflags="-X 'github.com/JLevconoks/k8ConsoleViewer/cmd.buildTime=$time' -X 'github.com/JLevconoks/k8ConsoleViewer/cmd.buildVersion=$version'" .
