+++ go.mod
@@ -5,6 +5,7 @@ go 1.13
 require (
 	github.com/gdamore/tcell v1.2.0
 	github.com/pkg/errors v0.8.1
+	github.com/spf13/cobra v0.0.5
 	k8s.io/api v0.0.0-20190531132109-d3f5f50bdd94
 	k8s.io/apimachinery v0.0.0-20190531131812-859a0ba5e71a
 	k8s.io/client-go v0.0.0-20190531132439-88ff0afc48bb
