--- version/version.go
@@ -1,13 +1,0 @@
-package version
-
-import "fmt"
-
-var (
-	BuildVersion string = ""
-	BuildTime    string = ""
-)
-
-func PrintVersion() {
-	fmt.Println("Version:", BuildVersion)
-	fmt.Println("Build Time:", BuildTime)
-}
